fixed editor layout
