+++ ui-packages/packages/form-details/src/envelope/components/FormEditor/tests/FormEditor.test.tsx
@@ -12,13 +12,29 @@ jest.mock('@patternfly/react-code-editor', () => ({
   }
 }));
 
+const formContent = {
+  name: 'from1',
+  source: {
+    'source-content': '<div><span>1</span></div>'
+  },
+  formConfiguration: {
+    resources: {
+      styles: {},
+      scripts: {}
+    },
+    schema:
+      '{"$schema":"http://json-schema.org/draft-07/schema#","type":"object","properties":{"candidate":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"salary":{"type":"integer"},"skills":{"type":"string"}},"input":true},"approve":{"type":"boolean","output":true}}}'
+  }
+};
 describe('FormView test', () => {
   it('render source - html', () => {
     const props = {
       code: '<div><span>1</span></div>',
       isSource: true,
       isConfig: false,
-      formType: 'html'
+      formType: 'html',
+      formContent: formContent,
+      setFormContent: jest.fn()
     };
     const wrapper = mount(<FormView {...props} />);
     expect(wrapper).toMatchSnapshot();
@@ -28,7 +44,9 @@ describe('FormView test', () => {
       code: '<React.FC><div><span>1</span></div></React.FC>',
       isSource: true,
       isConfig: false,
-      formType: 'tsx'
+      formType: 'tsx',
+      formContent: formContent,
+      setFormContent: jest.fn()
     };
     const wrapper = mount(<FormView {...props} />);
     expect(wrapper).toMatchSnapshot();
@@ -40,7 +58,9 @@ describe('FormView test', () => {
         2: '2'
       }),
       isSource: false,
-      isConfig: true
+      isConfig: true,
+      formContent: formContent,
+      setFormContent: jest.fn()
     };
     const wrapper = mount(<FormView {...props} />);
     expect(wrapper).toMatchSnapshot();
